test: improve test for showing updated qualities
Previous test made an assertion comparing two variables with similar origins--both originated from updatedItemsList.
Thus, this does not test whether or not a list of items is being updated as it should be.
The refactored test compares an items list that has been manipulated by the updateQuality and getHtmlListFromArray functions to produce a string of HTML with a hardcoded string representing what should be produced from the given items list.